requirements.txt
numpy>=1
pydub>=0.24.1
pyannote.audio>=3.1.0
torchvision>=0.17.0
transformers>=4.1.2
inquirer>=3.2.0
faster-whisper>=1.0.0
torch>=2.1.0